src/category_b/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
cassandra-driver==3.29.1
confluent-kafka==1.8.2
requests
kafka-python
